requirements.txt
pytesseract
Pillow
easyocr
numpy
